Apply suggested comments 1
