+++ consensus/src/units/mod.rs
@@ -8,8 +8,14 @@ use parking_lot::RwLock;
 use std::collections::HashMap;
 
 mod store;
+#[cfg(test)]
+mod testing;
 mod validator;
 pub(crate) use store::*;
+#[cfg(test)]
+pub use testing::{
+    add_units, create_units, creator_set, preunit_to_unchecked_signed_unit, preunit_to_unit,
+};
 pub use validator::{ValidationError, Validator};
 
 /// The coordinates of a unit, i.e. creator and round. In the absence of forks this uniquely
@@ -219,71 +225,16 @@ impl<H: Hasher> Unit<H> {
 }
 
 #[cfg(test)]
-pub use tests::{
-    add_units, create_units, creator_set, preunit_to_unchecked_signed_unit, preunit_to_unit,
-};
-
-#[cfg(test)]
 mod tests {
     use crate::{
-        creation::Creator as GenericCreator,
-        units::{
-            ControlHash, FullUnit as GenericFullUnit, PreUnit as GenericPreUnit,
-            UncheckedSignedUnit as GenericUncheckedSignedUnit, Unit as GenericUnit,
-        },
-        Hasher, NodeCount, NodeIndex, Round, SessionId, Signed,
+        units::{ControlHash, FullUnit as GenericFullUnit, PreUnit as GenericPreUnit},
+        Hasher, NodeIndex,
     };
-    use aleph_bft_mock::{Data, Hasher64, Keychain, Signature};
+    use aleph_bft_mock::{Data, Hasher64};
     use codec::{Decode, Encode};
 
-    type Creator = GenericCreator<Hasher64>;
     type PreUnit = GenericPreUnit<Hasher64>;
-    type Unit = GenericUnit<Hasher64>;
     type FullUnit = GenericFullUnit<Hasher64, Data>;
-    type UncheckedSignedUnit = GenericUncheckedSignedUnit<Hasher64, Data, Signature>;
-
-    pub fn creator_set(n_members: NodeCount) -> Vec<Creator> {
-        let mut result = Vec::new();
-        for i in 0..n_members.0 {
-            result.push(Creator::new(NodeIndex(i), n_members));
-        }
-        result
-    }
-
-    pub fn create_units<'a, C: Iterator<Item = &'a Creator>>(
-        creators: C,
-        round: Round,
-    ) -> Vec<(PreUnit, Vec<<Hasher64 as Hasher>::Hash>)> {
-        let mut result = Vec::new();
-        for creator in creators {
-            result.push(
-                creator
-                    .create_unit(round)
-                    .expect("Creation should succeed."),
-            );
-        }
-        result
-    }
-
-    pub fn preunit_to_unit(preunit: PreUnit, session_id: SessionId) -> Unit {
-        FullUnit::new(preunit, 0, session_id).unit()
-    }
-
-    pub fn add_units(creator: &mut Creator, units: &[Unit]) {
-        for unit in units {
-            creator.add_unit(unit);
-        }
-    }
-
-    pub async fn preunit_to_unchecked_signed_unit(
-        pu: PreUnit,
-        session_id: SessionId,
-        keybox: &Keychain,
-    ) -> UncheckedSignedUnit {
-        let full_unit = FullUnit::new(pu, 0, session_id);
-        let signed_unit = Signed::sign(full_unit, keybox).await;
-        signed_unit.into()
-    }
 
     #[test]
     fn test_full_unit_hash_is_correct() {
